+++ vortex-btrblocks/src/lib.rs
@@ -58,8 +58,6 @@ mod builder;
 mod canonical_compressor;
 /// Compression scheme implementations.
 pub mod schemes;
-#[cfg(feature = "parquet-variant")]
-pub mod variant;
 
 // Re-export framework types from vortex-compressor for backwards compatibility.
 // Btrblocks-specific exports.
